feat(grimoire): responsive two-pane shell, dense list panes, accent focus
Shell.svelte renders one two-pane frame at >=880px: a contextual index on
the left (dense single-column entity list or book section tree), the
reading surface (entity detail or chunk reader) on the right, so there is
no dead half-screen. Below 880px it collapses to the stacked master-detail
with back links. Pane widths live only in Shell; child routes render a
placeholder on desktop and the full list on mobile. The entity index is now
a dense name + type + grant-badge list, not a card grid. Topbar: the
viewpoint switcher gets breathing room off the right edge, and a single
accent :focus-visible ring replaces the default blue across the tree.
On-accent text routes through a new --grim-on-accent token.
